crosscontext-mcp/src/server.py
#!/usr/bin/env python3
"""
CrossContext MCP Server
Government Officer Context Engine with Trust & Safety Layer
"""

import asyncio
import os
from pathlib import Path

from fastmcp import FastMCP
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastMCP server
app = FastMCP("CrossContext MCP")

@app.tool()
async def echo_tool(message: str) -> str:
    """
    Simple echo tool for testing MCP connection.

    Args:
        message: The message to echo back

    Returns:
        The echoed message
    """
    return f"Echo: {message}"

@app.tool()
async def fetch_emails(query: str = "", person: str = "", date_range: str = "last_7_days", max_results: int = 10):
    """
    Fetch relevant emails from officer's inbox with Singapore government classification and PII redaction.

    Args:
        query: Search term or topic
        person: Filter by sender/recipient name/email
        date_range: Time range (last_7_days, last_30_days, today, etc.)
        max_results: Maximum number of emails to return

    Returns:
        Dict containing emails array with classification and redaction info
    """
    # Placeholder implementation - will be replaced with real Gmail API integration
    return {
        "emails": [
            {
                "id": "placeholder_1",
                "subject": f"Sample email about {query}",
                "from": "officer@gov.sg",
                "to": ["user@gov.sg"],
                "date": "2024-01-15T10:30:00Z",
                "snippet": "This is a sample email snippet with no PII.",
                "classification": "OFFICIAL (CLOSED)",
                "redacted": False
            }
        ],
        "total_count": 1,
        "audit_log_id": "audit_12345"
    }

@app.tool()
async def fetch_calendar(date_range: str = "today", meeting_title: str = "", attendee: str = "", include_past: bool = False):
    """
    Fetch officer's calendar events with Singapore government classification and PII redaction.

    Args:
        date_range: Time range (today, next_7_days, YYYY-MM-DD to YYYY-MM-DD)
        meeting_title: Search for specific meeting by title
        attendee: Filter by attendee name/email
        include_past: Whether to include past events

    Returns:
        Dict containing events array with classification and redaction info
    """
    # Placeholder implementation - will be replaced with real Google Calendar API integration
    return {
        "events": [
            {
                "id": "placeholder_event_1",
                "title": f"Meeting about {meeting_title}" if meeting_title else "Sample Meeting",
                "start_time": "2024-01-15T14:00:00Z",
                "end_time": "2024-01-15T15:00:00Z",
                "attendees": [
                    {
                        "name": "John Tan",
                        "email": "john.tan@gov.sg",
                        "role": "organizer"
                    }
                ],
                "location": "Conference Room A",
                "description": "Meeting description with no sensitive information.",
                "classification": "CONFIDENTIAL CLOUD-ELIGIBLE"
            }
        ],
        "audit_log_id": "audit_67890"
    }

@app.tool()
async def generate_briefing(context: str, briefing_type: str = "meeting_prep", emails_data: str = "[]", calendar_data: str = "[]", policy_data: str = "[]", stakeholder_data: str = "{}"):
    """
    Synthesize briefing from pre-fetched data sources with Singapore government classification.

    Args:
        context: Meeting topic or decision context
        briefing_type: Type of briefing ('meeting_prep' | 'decision_memo' | 'stakeholder_brief')
        emails_data: JSON string of pre-fetched email data
        calendar_data: JSON string of pre-fetched calendar data
        policy_data: JSON string of pre-fetched policy data
        stakeholder_data: JSON string of pre-fetched stakeholder data

    Returns:
        Structured briefing with citations and highest classification level
    """
    # Placeholder implementation - will be enhanced with LLM synthesis
    return {
        "briefing": {
            "title": f"Briefing: {context}",
            "summary": f"Executive summary for {briefing_type} on {context}",
            "key_points": [
                "Key point 1 from synthesized data",
                "Key point 2 from stakeholder context",
                "Key point 3 from policy review"
            ],
            "background": "Contextual information from multiple sources",
            "stakeholder_analysis": "Analysis of involved parties",
            "recommendations": "Suggested talking points and actions",
            "citations": [
                {
                    "source_type": "email",
                    "excerpt": "Relevant email excerpt",
                    "classification": "OFFICIAL (CLOSED)"
                }
            ],
            "highest_classification": "CONFIDENTIAL CLOUD-ELIGIBLE"
        },
        "audit_log_id": "audit_brief_12345"
    }

if __name__ == "__main__":
    # Run the MCP server
    import asyncio
    asyncio.run(app.run_stdio_async())
